backend/english_learning_app/requirements.txt
aioredis==1.3.1
ansicon==1.89.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.7.2
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
blessed==1.21.0
certifi==2025.4.26
cffi==1.17.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.5
daphne==3.0.2
distlib==0.3.8
Django==4.2.4
django-cors-headers==4.6.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-picklefield==3.3
django-q==1.3.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
hiredis==3.2.1
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jinxed==1.3.0
Markdown==3.7
msgpack==1.1.1
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==3.5.3
requests==2.32.3
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.4.4
Twisted==25.5.0
txaio==25.6.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.14.1
tzdata==2023.3
urllib3==2.4.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.26.3
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
zope.interface==7.2
drf_extra_fields==3.7.0